feat: ✨ get input going
